synth_recorder_gui/requirements.txt
mido==1.3.3
pygame-ce==2.5.3
pygame_gui==0.6.13
pyserial==3.5